# swagger.yaml
openapi: 3.0.3
info:
  title: Grocery Ordering Platform API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
  /api/user/profile:
    get:
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  email: { type: string }
                  role: { type: string }
  /api/product:
    get:
      summary: List all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    name: { type: string }
                    description: { type: string }
                    unit: { type: string }
                    price: { type: number }
                    stockQuantity: { type: integer }
  /api/product/{id}:
    get:
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Product details
  /api/cart:
    post:
      summary: Add item to cart
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { type: string }
                productId: { type: integer }
                quantity: { type: integer }
      responses:
        '200':
          description: Cart item added
  /api/cart/{userId}:
    get:
      summary: View cart
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cart items
    delete:
      summary: Clear cart
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cart cleared
  /api/order/{userId}:
    post:
      summary: Place a new order
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order placed
    get:
      summary: View past orders
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of orders
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT